# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Group": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "Group",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "TransactionCreate": {
        "properties": {
          "amount": {
            "description": "Total amount in smallest currency unit (e.g., cents)",
            "title": "Amount",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "group_id": {
            "format": "uuid",
            "title": "Group Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/TransactionParticipantCreate"
            },
            "title": "Participants",
            "type": "array"
          },
          "payer_id": {
            "format": "uuid",
            "title": "Payer Id",
            "type": "string"
          },
          "purchased_on": {
            "format": "date-time",
            "title": "Purchased On",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transaction_type": {
            "$ref": "#/components/schemas/TransactionType",
            "default": "EVEN"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "title",
          "group_id",
          "payer_id",
          "participants"
        ],
        "title": "TransactionCreate",
        "type": "object"
      },
      "TransactionParticipantCreate": {
        "properties": {
          "amount_owed": {
            "title": "Amount Owed",
            "type": "integer"
          },
          "debtor_id": {
            "format": "uuid",
            "title": "Debtor Id",
            "type": "string"
          }
        },
        "required": [
          "amount_owed",
          "debtor_id"
        ],
        "title": "TransactionParticipantCreate",
        "type": "object"
      },
      "TransactionParticipantRead": {
        "properties": {
          "amount_owed": {
            "title": "Amount Owed",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "debtor": {
            "$ref": "#/components/schemas/User"
          },
          "debtor_id": {
            "format": "uuid",
            "title": "Debtor Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "transaction_id": {
            "format": "uuid",
            "title": "Transaction Id",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "amount_owed",
          "transaction_id",
          "debtor_id",
          "id",
          "created_at",
          "updated_at",
          "debtor"
        ],
        "title": "TransactionParticipantRead",
        "type": "object"
      },
      "TransactionRead": {
        "properties": {
          "amount": {
            "description": "Total amount in smallest currency unit (e.g., cents)",
            "title": "Amount",
            "type": "integer"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "group": {
            "$ref": "#/components/schemas/Group"
          },
          "group_id": {
            "format": "uuid",
            "title": "Group Id",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "participants": {
            "items": {
              "$ref": "#/components/schemas/TransactionParticipantRead"
            },
            "title": "Participants",
            "type": "array"
          },
          "payer": {
            "$ref": "#/components/schemas/User"
          },
          "payer_id": {
            "format": "uuid",
            "title": "Payer Id",
            "type": "string"
          },
          "purchased_on": {
            "format": "date-time",
            "title": "Purchased On",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "transaction_type": {
            "$ref": "#/components/schemas/TransactionType",
            "default": "EVEN"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "amount",
          "title",
          "group_id",
          "payer_id",
          "participants",
          "payer",
          "group"
        ],
        "title": "TransactionRead",
        "type": "object"
      },
      "TransactionType": {
        "enum": [
          "EVEN",
          "AMOUNT",
          "PERCENTAGE"
        ],
        "title": "TransactionType",
        "type": "string"
      },
      "TransactionUpdate": {
        "properties": {
          "amount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Amount"
          },
          "payer_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payer Id"
          },
          "purchased_on": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Purchased On"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "transaction_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransactionType"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "TransactionUpdate",
        "type": "object"
      },
...
      "User": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "email": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Email",
            "type": "string"
          },
          "email_verification_token": {
            "title": "Email Verification Token",
            "type": "integer"
          },
          "email_verified": {
            "default": false,
            "title": "Email Verified",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "password": {
            "minLength": 8,
            "title": "Password",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "username": {
            "maxLength": 50,
            "minLength": 1,
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "email",
          "password"
        ],
        "title": "User",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "FastAPI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/groups/{group_id}/transactions": {
      "get": {
        "operationId": "read_group_transactions_groups__group_id__transactions_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "format": "uuid",
                  "type": "string"
                }
              ],
              "title": "Group Id"
            }
          },
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionRead"
                  },
                  "title": "Response Read Group Transactions Groups  Group Id  Transactions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Group Transactions",
        "tags": [
          "groups",
          "transactions"
        ]
      }
    },
...
    "/transactions/": {
      "get": {
        "operationId": "read_transactions_user_is_participant_in_transactions__get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "group_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Group Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionRead"
                  },
                  "title": "Response Read Transactions User Is Participant In Transactions  Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Transactions User Is Participant In",
        "tags": [
          "transactions"
        ]
      },
      "post": {
        "operationId": "create_transaction_transactions__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Transaction",
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/{transaction_id}": {
      "delete": {
        "operationId": "delete_transaction_transactions__transaction_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transaction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Transaction",
        "tags": [
          "transactions"
        ]
      },
      "get": {
        "operationId": "read_transaction_transactions__transaction_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transaction Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Read Transaction",
        "tags": [
          "transactions"
        ]
      },
      "put": {
        "operationId": "update_transaction_transactions__transaction_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "transaction_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Transaction Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransactionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "Not found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Transaction",
        "tags": [
          "transactions"
        ]
      }
    }
  }
}